Extract init mocks in a reusable file
